240220 6th commit : pid의 타입은 int가 아닌 long

+++ stock/src/main/java/com/miniproject/stock/service/StockService.java
@@ -6,7 +6,7 @@ import com.miniproject.stock.entity.Stock;
 
 public interface StockService {
 
-    public Stock search(int pid);
+    public Stock search(long pid);
 
     public boolean insert(Stock stock);
 
@@ -14,8 +14,8 @@ public interface StockService {
     
     public List<Stock> selectAll();
 
-    public boolean updateStock(int pid, int num); // 입고 시 양수, 출고 시 음수 전달
+    public boolean updateStock(long pid, int num); // 입고 시 양수, 출고 시 음수 전달
     
-    public boolean updatePrice(int pid, int price); // 변경할 가격 전달
+    public boolean updatePrice(long pid, int price); // 변경할 가격 전달
 
 }
